@@ -5,7 +5,7 @@ describe MailRoom::Configuration do
 
   describe 'set_mailboxes' do
     context 'with config_path' do
-      let(:configuration) { MailRoom::Configuration.new(:config_path => config_path) }
+      let(:configuration) { MailRoom::Configuration.new(config_path: config_path) }
 
       it 'parses yaml into mailbox objects' do
         MailRoom::Mailbox.stubs(:new).returns('mailbox1', 'mailbox2')
@@ -19,10 +19,9 @@ describe MailRoom::Configuration do
 
       it 'sets mailboxes to an empty set' do
         MailRoom::Mailbox.stubs(:new)
+        MailRoom::Mailbox.expects(:new).never
 
         expect(configuration.mailboxes).to eq([])
-
-        expect(MailRoom::Mailbox).to have_received(:new).never
       end
     end
   end
