@@ -1,19 +1,16 @@
+require 'date'
 
 module MailRoom
   class CrashHandler
-
-    attr_reader :error, :format
-
     SUPPORTED_FORMATS = %w[json none]
 
-    def initialize(error:, format:)
-      @error = error
-      @format = format
+    def initialize(stream=STDOUT)
+      @stream = stream
     end
 
-    def handle
+    def handle(error, format)
       if format == 'json'
-        puts json
+        @stream.puts json(error)
         return
       end
 
@@ -22,8 +19,8 @@ module MailRoom
 
     private
 
-    def json
-      { time: Time.now, severity: :fatal, message: error.message, backtrace: error.backtrace }.to_json
+    def json(error)
+      { time: DateTime.now.iso8601(3), severity: :fatal, message: error.message, backtrace: error.backtrace }.to_json
     end
   end
 end
