@@ -1,3 +1,4 @@
+require 'date'
 require 'logger'
 require 'json'
 
@@ -10,11 +11,24 @@ module MailRoom
 
         data = {}
         data[:severity] = severity
-        data[:time] = timestamp || Time.now.to_s
+        data[:time] = format_timestamp(timestamp || Time.now)
         # only accept a Hash
         data.merge!(message)
 
         data.to_json + "\n"
+      end
+
+      private
+
+      def format_timestamp(timestamp)
+        case timestamp
+        when Time
+          timestamp.to_datetime.iso8601(3).to_s
+        when DateTime
+          timestamp.iso8601(3).to_s
+        else
+          timestamp
+        end
       end
     end
   end
