@@ -3,14 +3,12 @@ require 'spec_helper'
 describe MailRoom::Coordinator do
   describe '#initialize' do
     it 'builds a watcher for each mailbox' do
-      MailRoom::MailboxWatcher.stubs(:new).returns('watcher1', 'watcher2')
+      MailRoom::MailboxWatcher.expects(:new).with('mailbox1').returns('watcher1')
+      MailRoom::MailboxWatcher.expects(:new).with('mailbox2').returns('watcher2')
 
       coordinator = MailRoom::Coordinator.new(['mailbox1', 'mailbox2'])
 
       expect(coordinator.watchers).to eq(['watcher1', 'watcher2'])
-
-      expect(MailRoom::MailboxWatcher).to have_received(:new).with('mailbox1')
-      expect(MailRoom::MailboxWatcher).to have_received(:new).with('mailbox2')
     end
 
     it 'makes no watchers when mailboxes is empty' do
@@ -27,24 +25,27 @@ describe MailRoom::Coordinator do
       MailRoom::MailboxWatcher.stubs(:new).returns(watcher)
       coordinator = MailRoom::Coordinator.new(['mailbox1'])
       coordinator.stubs(:sleep_while_running)
+      watcher.expects(:run)
+      watcher.expects(:quit)
+
       coordinator.run
-      expect(watcher).to have_received(:run)
-      expect(watcher).to have_received(:quit)
     end
     
     it 'should go to sleep after running watchers' do
       coordinator = MailRoom::Coordinator.new([])
       coordinator.stubs(:running=)
       coordinator.stubs(:running?).returns(false)
+      coordinator.expects(:running=).with(true)
+      coordinator.expects(:running?)
+
       coordinator.run
-      expect(coordinator).to have_received(:running=).with(true)
-      expect(coordinator).to have_received(:running?)
     end
 
     it 'should set attribute running to true' do
       coordinator = MailRoom::Coordinator.new([])
       coordinator.stubs(:sleep_while_running)
       coordinator.run
+
       expect(coordinator.running).to eq(true)
     end
   end
@@ -54,8 +55,9 @@ describe MailRoom::Coordinator do
       watcher = stub(:quit)
       MailRoom::MailboxWatcher.stubs(:new).returns(watcher)
       coordinator = MailRoom::Coordinator.new(['mailbox1'])
+      watcher.expects(:quit)
+
       coordinator.quit
-      expect(watcher).to have_received(:quit)
     end
   end
 end
