@@ -4,21 +4,22 @@ describe MailRoom::CrashHandler do
 
   let(:error_message) { "oh noes!" }
   let(:error) { RuntimeError.new(error_message) }
+  let(:stdout) { StringIO.new }
 
   describe '#handle' do
 
-    subject{ described_class.new(error: error, format: format) }
+    subject{ described_class.new(stdout).handle(error, format) }
 
     context 'when given a json format' do
       let(:format) { 'json' }
-      let(:fake_json) do
-        { message: error_message }.to_json
-      end
 
-      it 'outputs the result of json to stdout' do
-        subject.stubs(:json).returns(fake_json)
+      it 'writes a json message to stdout' do
+        subject
+        stdout.rewind
+        output = stdout.read
 
-        expect{ subject.handle }.to output(/\"message\":\"#{error_message}\"/).to_stdout
+        expect(output).to end_with("\n")
+        expect(JSON.parse(output)['message']).to eq(error_message)
       end
     end
 
@@ -26,7 +27,7 @@ describe MailRoom::CrashHandler do
       let(:format) { "" }
 
       it 'raises an error as designed' do
-        expect{ subject.handle }.to raise_error(error.class, error_message)
+        expect{ subject }.to raise_error(error.class, error_message)
       end
     end
 
@@ -34,7 +35,7 @@ describe MailRoom::CrashHandler do
       let(:format) { "nonsense" }
 
       it 'raises an error as designed' do
-        expect{ subject.handle }.to raise_error(error.class, error_message)
+        expect{ subject }.to raise_error(error.class, error_message)
       end
     end
   end
