@@ -18,20 +18,15 @@ describe MailRoom::Delivery::Que do
     let(:options) {MailRoom::Delivery::Que::Options.new(mailbox)}
 
     it 'stores the message in que_jobs table' do
-      PG.stubs(:connect).returns(connection)
-      connection.stubs(:exec)
-
-      MailRoom::Delivery::Que.new(options).deliver('email')
-
-      expect(PG).to have_received(:connect).with({
+      PG.expects(:connect).with({
         host: 'localhost',
         port: 5432,
         dbname: 'delivery_test',
         user: 'postgres',
         password: ''
-      })
+      }).returns(connection)
 
-      expect(connection).to have_received(:exec).with(
+      connection.expects(:exec).with(
         "INSERT INTO que_jobs (priority, job_class, queue, args) VALUES ($1, $2, $3, $4)",
         [
           5,
@@ -40,6 +35,8 @@ describe MailRoom::Delivery::Que do
           JSON.dump(['email'])
         ]
       )
+
+      MailRoom::Delivery::Que.new(options).deliver('email')
     end
   end
 end
