@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe MailRoom::Connection do
+describe MailRoom::IMAP::Connection do
   let(:imap) {stub}
   let(:mailbox) {build_mailbox(delete_after_delivery: true, expunge_deleted: true)}
 
@@ -9,7 +9,9 @@ describe MailRoom::Connection do
   end
 
   context "with imap set up" do
-    let(:connection) {MailRoom::Connection.new(mailbox)}
+    let(:connection) {MailRoom::IMAP::Connection.new(mailbox)}
+    let(:uid) { 1 }
+    let(:seqno) { 8 }
 
     before :each do
       imap.stubs(:starttls)
@@ -36,30 +38,24 @@ describe MailRoom::Connection do
     end
 
     it "waits for a message to process" do
-      new_message = 'a message'
-      new_message.stubs(:seqno).returns(8)
+      new_message = MailRoom::IMAP::Message.new(uid: uid, body: 'a message', seqno: seqno)
 
       connection.on_new_message do |message|
         expect(message).to eq(new_message)
         true
       end
 
-      mailbox.stubs(:deliver?).returns(true)
+      attr = { 'UID' => uid, 'RFC822' => new_message.body }
+      fetch_data = Net::IMAP::FetchData.new(seqno, attr)
 
-      imap.stubs(:idle)
-      imap.stubs(:uid_search).returns([]).then.returns([1])
-      imap.stubs(:uid_fetch).returns([new_message])
-      imap.stubs(:store)
-      imap.stubs(:expunge)
+      imap.expects(:idle)
+      imap.stubs(:uid_search).with(mailbox.search_command).returns([], [uid])
+      imap.expects(:uid_fetch).with([uid], "RFC822").returns([fetch_data])
+      mailbox.expects(:deliver?).with(uid).returns(true)
+      imap.expects(:store).with(seqno, "+FLAGS", [Net::IMAP::DELETED])
+      imap.expects(:expunge).once
 
       connection.wait
-
-      expect(imap).to have_received(:idle)
-      expect(imap).to have_received(:uid_search).with(mailbox.search_command).twice
-      expect(imap).to have_received(:uid_fetch).with([1], "RFC822")
-      expect(mailbox).to have_received(:deliver?).with(1)
-      expect(imap).to have_received(:store).with(8, "+FLAGS", [Net::IMAP::DELETED])
-      expect(imap).to have_received(:expunge).once
     end
   end
 end
